+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+- Improved support for stateless live components. They don't store mock state objects in Redis anymore.
+
 ## 1.8.0 (2023-12-18)
 
 - Added helper method StateAddress.with_component_id() to create a StateAddress with the same session ID, but a different component ID.